fix: revert change to syncFunctionInner callback.
Notes: https://github.com/nrkno/tv-automation-server-core/pull/262#discussion_r453453481
Nytamin: I'm sure i had a motivation for it when I wrote it, but I can't remember exactly where the issue was, and I seem to not be able to reproduce it now..
I'm going to pull this out for now, but leave this comment for posterity, in case the issue shows up again:
If I remember correctly, I had an issue where the last argument was undefined for some reason. If I remember correctly (and I don't think I do, because I wasn't able to reproduce it now) the Meteor.wrapAsync was able to handle it so I made this to be as analogous to that as possible.