feat: Gui Optimizations (#327)
* fix: make getResolvedSegment get executed less often

The tracker for all visible segments was re-running the autorun on every
take because it was reacting to current and next part changing.
Now the current and next parts affecting a segment will come as props.

* fix: make getResolvedSegment get executed even less often

The tracker was re-running the autorun when timings of Pieces were changing.

* refactor: move some calculations out of getResolvedSegment

* fix: properly handle setting a part as next

Props can update faster than the durations so we need to account for that.

* fix: not scrolling to a manually set segment on take

* fix: segments not rewound when resetting the rundown

Resetting the livePosition seems to fix the problem of segments
occasionally not rewinding correctly when resetting the rundown.
onFollowLiveLine called in componentDidUpdate was setting wrong scrollLeft based on old livePosition.

* fix: segments not autorewinding when leaving

State is updated in componentDidUpdate later than onRewindSegment is
called. this.isLiveSegment mirrors this.state.isLiveSegment but is set earlier.

* chore: remove a fix for IntersectionObserver bug in Chrome 76

The bug seems to be resolved by now, so the workaround could be removed.

* fix: a problem with double scrolling after a take

* chore: make the set as next menu item disabled

The item was not displayed when the context part was dynamically inserted. This has been replaced with displaying a disabled menu item, like with the current part.

* fix: prevent viewport jumping on takes

* chore: remove leftover console.log

There was a console.log in SegmentTimelineContainer left over after doing some investigations

* feat: an attempt at fixing the part jumping back-forth issue

* chore: Lint

* fix: restore a flush of temporaryPartInstances in RundownTiming

* fix: make sure we don't treat scroll promise rejections as errors to be printed to console wholesale

* chore: lint

* chore: fix a bad merge in client/lib/rundown.ts

* fix: remove unused field modifier
This PR has been retargeted towards R25 which removed some properties from Pieces

Co-authored-by: ianshade <zegzulakrzysztof@gmail.com>
Co-authored-by: Jan Starzak <jan.starzak@gmail.com>