fix: selective copying of existing part properties during ingest
Before PartInstances, we had to preserve many Part properties as they would be added during playout. Playout no longer modifies them, so we only have to cater to any other ingest property flows (mos commands). This changes from a blocklist to an allowlist for the properties to avoid preserving properties we shouldnt