fix: disable/undo disable piece fails or throws an error
There have been a couple of issues with Disable/Undo Disable functionality. First of all, the GUI would not show disabled pieces, because shared backend-fronted methods used by getResolvedSegment() for processing Piece Instances would filter them out, because backend doesn't care about disabled pieces. Then there was a problem that meant that in some conditions, when trying to undo-disable a Piece, the Piece would not be found, because the method would end too early due to it breaking out of a loop too quickly. Finally, there was an additional problem with an error being printed out if no piece do be disabled was found, because a cache was created, but - since no modification was done - it wasn't saved back.