feat: initial implementation of overflow content countdown
This creates a countdown attached to the On Air line, if the piece underneath it has some overflowing content/ends before the end of the part. If the end of the clip is at the same position as the end of the part, the counter shouldn't show. This is just an initial implementation. Needs more testing with actual use scenarios.