feat: piece freeze frame countdown (#426)
* fix(SplitsSourceRenderer): don't print label widths out to console

* feat: initial implementation of overflow content countdown

This creates a countdown attached to the On Air line, if the piece underneath it has some overflowing content/ends before the end of the part. If the end of the clip is at the same position as the end of the part, the counter shouldn't show. This is just an initial implementation. Needs more testing with actual use scenarios.

* fix: Piece countdown label placement

The Piece countdown wasn't placed correctly, because a CSS variable wasn't defined for the correct scope

* fix: overlay screen layout when an upcoming rundown is present

* fix: do not show the counter after the onAir line has passed the Piece

* chore: typos and simplify code