fix(Prompter): If collection changes are slow, Prompter may receive incomplete updates, and render some Script missing
In non-ideal network situations, it's possible that the Piece/PieceInstance collection updates will not stream at once into the Prompter. If that takes more than 250 ms, then the prompter will update once (without a Piece of script) and then once again (with a Piece of script). This change introduces an alternative timer that, if a Script Piece is not present in an update, a 2s delay will wait for that Script Piece to appear, before rendering it gone.