feat: move Rundown._rank to RundownPlaylist.rundownIdsInOrder
This reduces the complexity of reordering the rundowns in the playlist, and allows us to easily avoid some race conditions that could occur.
The rank of the rundowns has long been decided by the blueprints at the time of generating the playlist, so conceptually the data has belonged to the playlist for a while. This enforces that ownership in the data model