fix: Improve clock accuracy
Instead of using setInterval to get a tick every second, starting at a
random point an slowly drifting a ms or 2 every few updates, set a 
timeout for the number of ms until the next update is due.

For example, if we have an interval of 1000ms (1s), and we start at
12:00:00.123, we will set a timeout of 877ms so that the next update
should be at exactly 12:00:01.000.

In practice, depending on how busy the browser’s event loop is etc. it
might be at e.g. 12:00:01.002, in which case the next update will set a timeout of 998ms.